% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_function.R
\name{filter2}
\alias{filter2}
\title{Title: Filter your rows}
\usage{
filter2(data, vector)
}
\arguments{
\item{vector}{A numeric vector indicating which rows to keep}

\item{data.}{A dataframe}
}
\value{
a new database with filtered rows
}
\description{
Title: Filter your rows
}
\examples{
filter2(iris, 2)
}
