% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_function.R
\name{select2}
\alias{select2}
\title{Title  Select columns from database}
\usage{
select2(data, vector)
}
\arguments{
\item{data}{a dataframe}

\item{vector}{columns to select}
}
\value{
a new dataset with the columns selected
}
\description{
Title  Select columns from database
}
\examples{
select2(iris, "Sepal.Length")
}
